Don't laugh, this might actually be a good idea.
I keep an eye on my line lengths, 80 is good guidance, and I try not to
go above it... a lot ;-)
Still, when typing a longer line, I've catched myself keeping an eye on
the column counter in the statusline.  This is ridiculous; so instead of
distracting myself, why not have it visually 'in-your-face' when
editing?  I always thought that colored vertical line was an ugly
distraction, but Jelmer might be very right about this one.