Replacing Powerline with Airline.
Powerline is being rewritten in Python, I'm not convinced (yet?) that
that is a good idea: seems quite a big dependency, sounds troublesome
when I'm on a windoos machine (unlikely, but still).  I'm also worried
about loading time.
Therefor, I'm giving airline a try (thx 2 @davelens who pointed me to
it), but it might be that over time I'll switch back to plain
statusline= in vimrc, we'll see...