Don't use <cr> as mapping key in normal mode.
It interferes with behaviour in buffers like 'q:' or 'Ack' results; <cr>
in those buffers act directly on the line you are on, but since <cr> was
mapped to opening a line, this would not work correctly.
By adding the CONTROL- and SHIFT key to the mapping, this works
correctly again, albeit we'll have to retrain the muscle memory.