CtrlP now fuzzy search throughout project.
I don't know what the exact reason was; either an update in CtrlP or the
fact that I'm doing a lot larger projects (as in: subdirs up the wazoo),
but somewhere, somehow CtrlP did not fuzzy matched all the files in a
project.

At first, I thought the subdirectory level was too deep, but after some
fiddling about I found out that it was due to the max search file
setting, ```g:ctrlp_max_files```
Setting this to '0' fixed everything, back to happy now.