Remove colorscheme 'perzen' (and 'zenburn' on which it was based).
I've been working continuously in Solarized light for years now without
looking back to the zenburn derivate I had been using for years before
that.  So I'm removing zenburn/perzen from my repository; if I need a
darker (or different) colorscheme, it'll probably be for just a short
period of time (ie. < 1 day) and so it doesn't matter that much which
color scheme I'll use then (and Solarized does provide a dark version).