Remove :nohlsearch keymap, <space> is now <leader> key.
Having the space bar as both leader key and shortcut to :nohlsearch
makes that it takes a split second (second key timeout) for the
highlighted search to disappear.  I can type `:noh<tab>` about as quick.
Further more, this is one custom key mapping less and thus less
frustration when editing on another machine with a default vim (using
sensible.vim or not).