Do _not_ jump to the last known cursor position in a commit message.
I got this from the defaults.vim that is now shipped with Vim 8.0.
I had to replicate the augroup because other wise Vim complained about
the !~# = 'commit' part (I think, it was also highlighted differently
when in and out of the autcommand group).
I'm not too sure what is going on, but I'm already glad the cursor will
be top left when git committing.

Closes #16