Add modeline to .vimrc since we have reinstated default foldmethod.
By default, the foldmethod in Vim is "manual".  I had foldmethod=marker
for a very long time in vimrc, but have started to find that this is
more of a file (buffer) specific setting, so I deleted it.  This _does_
mean that I need to set this specifically for the .vimrc file.