.vimrc rebrand and restyle.
I took a peek at @spf13's .vimrc (https://github.com/spf13/spf13-vim/blob/3.0/.vimrc)
and like the way comments + fold markers are used as an integral part of
the .vimrc language.  It's "simple" in the most positive sense of the
word.
Since I installed figlet yesterday and played with it in [fdd0636], this
was also the ideal moment to rebrand my Vim configuration... "DubVim"...
It's a working title I guess.