Remove <leader>r for editing vimrc.
It was taken by PHP Refactor and I'm not willing to change its default
key mapping.
I'm not going to redefine a key map for editing my vimrc.  I revisit
vimrc once in a while and _then_ a key map could be nice, but it's not
something I need to have mapped permanently.

Closes #20