Add mapping to quickly change to a project directory.
Projects are often prefixed with some identifiers to denote some extra
information on the project itself, for instance: "drupal8_site_<name>"
to indicate that project <name> is a Drupal 8 site.
Since I clone projects like that all the time, without renaming the
directories locally, this makes it more difficult to switch to the
relevant directory.  With this mapping I just have to type <Leader>e,
then part of the name of the project and hit <Tab>.