Adhere to the 80 col cursor column indicator.
Since [6731f7] the indent level has gone up in the vimrc.  That's
actually a good thing in this case, but it does mean that some line
cross the 80-col soft limit that I usually have for code.

Where immediately possible, I've re-formatted the code, except for that
few lines that I need to think about how I'm going to split them up.