Forcing Python 2.
Something was triggering Vim into Python3 before Vdebug got loaded
resulting in Vdebug not loading at all (it needs Python 2, it will not
work on Python3).
I tried to figure some stuff out, but couldn't directly pinpoint where
Py3 was set.  Fortunately, Twitter to the rescue:
https://twitter.com/chrisheithoff/status/928259481930153984
which is sub optimal, but it _does_ work and it is a very simple,
pragmatic solution...