Ignore dirty state of php-refactoring-toolbox.
This plugin doesn't have doc/tags in its .gitignore, which makes that
when the help tags are generated, the directory of
php-refactoring-toolbox appears to have changed.  We could add these
tags files to our .gitignore, but ignoring the dirty state of submodules
is fine enough as far as I am concerned.