Remove plugin webapi.
IIRC this was installed as a dependency for Gist plugin, but since this
has also moved, I'm deleting webapi also.  There are some good things in
there, like http url en/decoding, but there are other ways to accomplish
this; Tpope has a plugin that also does this and I can always shell out
to some command or even steal the string manipulation parts without
slurping in all the webapi related parts.