Switch to IBM Plex Mono font.
In [4413dc969c63f] I switchted to Consolas, which I like a lot, but as
mentioned in that commit, the installation is suboptimal due to the
Microsoft apps I needed to have installed.
IBM Plex also seems very nice, and the fact that you can just install it
via homebrew is big plus.