Use file#ShortPath() to display filename in LightLine.
The default way that LightLine shows the name of the file in the current
buffer is confusing when doing a `:Gdiff` because you cannot distinguish
immediately between the normal file and the version in the `.git/`
directory.
By using `file#ShortPath` we can fix this, but by default it did not
show the filename itself, which is now changed... and thus the old
`statusline` config (still there for doc purposes or for when I'm tired
of LightLine) needs to change too.