Remove keymap for ToggleWritingMode command.
Switching back and forth was ok when testing stuff out, but in reality
I'd switch to writing mode to do some writing, so I'll be in that mode
for quite some time, not warranting a separate <leader>-key mapping.

Furthermore, the 'w' letter is already in muscle memory in Bash for
different action (to 'w'ork on a different project), which would tax
muscle memory too much or confuse me.