Put Gruvbox in low contrast mode, whether light or dark background.
This is easier on the eyes, and it is easily changeable to "medium" (the
default value) or "hard" (more suitable for sunny conditions).

I'm thinking about putting the color scheme choice and -settings into a
separate dubvim/....vim file, but for now, I'm doing it this way;
ensuring the contrast global variable is set before choosing colorscheme
and background.