Disable Ale by default.
I'm using Ale only when working in a larger project, not when doing
small (one-off) scripts etc...  It is a bit annoying that Ale shows
linting errors (uses PEAR style by default for PHP for instance :sigh:)
when doing some small scripting.  Therefor, I've disabled it by
default... when all is set up in a project, it is easily (re)activated
by `:ALEEnable`.