Do use showcmd for gauging the size of a visual selection.
In [a3c5e909] `noshowcmd` was set because I never had a use for showing
the commands executed... but this also means that the size of a visual
selection is not shown (number of selected characters in visual mode, or
number of lines in vertical/line visual mode).  And that _is_ a useful
feature, so I'm re-enabling it.