Remove TagBar, not using it that much and it may die.
There was an update in the Tagbar plugin stating that the original
author was looking for new maintainers.  This might turn out well, and I
hope for the daily Tagbar users that it does, but since I hardly ever
use it, relying on FZF's tags possibilities, it feels I'm better off by
just removing it from my Vim configuration.