Do not show cursor line in help screens.
The cursorline is handy, but in help files it is a distraction, so we
check in what type of buffer we are before we show/hide the cursor line.
Apparently we need to add the BufEnter autocmd event for this to work,
otherwise the cursor line will not be hidden the first time a help
screen is requested.

Closes #41.