Explicit set pyunit as the compiler for Test.
This is, for the time being, the only way I can have the file + line
number in the quick fix window parsed correctly so I can navigate to it.
The default-run-out-of-the-box provided the correct test output, but
failed to have the file and line number 'clickable', I'm not too sure
what's causing it.  Might even be due to my structure of test suite in
the project where I'm using this now... might also be some bug in Test,
even in Dispatch (although, I could reproduce the erroneous behavior by
removing the Dispatch plugin completely).

Fixes #48.