Do not show FZF preview when window is not wide enough.
In [f6109e958da80379287af7f490f7514f8268823e] I've made the FZF preview
always visible, but this is too cluttered and unreadable with long paths
and Vim taking only half a laptop screen.  It is more important to get
more of the file path than it is to see a preview.