Ale and VDebug are removed.
VDebug: Hardly ever used it, only in the beginning for its fanciness,
but I've ran in too much fiddling with it and I prefer testing/TDD and,
if necessary print_r(); exit; and the likes...

Ale: appeared to be unstable at times, ie; it reported correct lint
messages but on the wrong line numbers which was very confusing.  It
also didn't work 'out of the box' for the sheer size of the plugin.
I might replace with a more lightweight syntactic, or something custom
roled via vim-test/vim-dispatch.