FZF: Preview window can be toggled, not shown by default.
We hide the preview window by default, but can be toggled by the <c-/>
keystroke.
Since previews were proven not to be _that_ helpful after all, it seemed
bes to hide it by default.  Also, instead of to the default right of the
FZF screen, it is now shown above the fuzzy finder, showing more width,
which comes in handy in smaller screens (like in a 50/50 Vim/Shell
situation).

Closes #52