Smoother QuickFix list navigation
Previous qf navigation bindings were a bit slow due to keyboard timers
being set because of the use of double-pressing the same key (For
instance `<Left><Left>` to go to the previous _file_ in the quickfix
list.
We don't do that now, instead using a SHIFT-combination.  At the same
time I've swapped out the LEFT/RIGHT for UP/DOWN respectively because
that makes more sense visually.  Which frees up `<Left>` and `<Right>`
to go to the outlier ends of the list.

Closes #54.