vim-plug now manages plugins.
In using the git submodules approach, I've stumbled upon merge conflict
issues a couple of times, which put me off enough that I wanted to move
to a plugin manager.
vim-plug seems to be well maintained and light weight enough to my
likings.  I've had an issue with not fully updating/installing
automatically, but nothing that a manual `:PlugInstall` with a reload of
Neovim can't fix, so I opted to just echo a reminder and leave it at
that.

Added benefit is that I do not need to maintain git submodules, nor do I
need the `update` and `install` shell scripts; this can be done from
within Vim.
It also gets rid of all the "Plugin update:..." git commits.

Might be that I need to tweak/alter finetune some things, but for the
moment, alls seems to be working just fine.

Closes #30