Back to Solarized Light!
After years of enjoying Gruvbox, I've returned to Solarized.  I gave it
a try because I was looking at the colorschemes on
https://github.com/lifepillar, especially gruvbox8 and solarized8 and
when I switched to Solarized 8, I was immediately struck by its
readability, it seemed to be less effort, almost everything was
noticeably clearer, the only thing that was a bit less apparent is the
horizontal split, which isn't clear when a closed fold is pressed
against the status line.  But that's about it really.
Especially when using the 'text' font face of IBM Plex Mono font in
iTerm, this is so nice to read!