Changed (for the better!) FZF mappings
The :GFiles mapping in [f7cd666] was a good idea, but not optimal:
In the beginning of a project (and less frequently afterwards) it
happens that I wanted to open a file not yet committed.  The poorly
chosen <Leader>E provide for this, but had too many keys to press -> 'o'
for 'Open' seems a lot better.

Then it occurred to me that I was using the help system more than
searching in the file history.

And last but not least; searching and _finding_ stuff in a project is
important, so <Leader>f has been augmented for that, while searching in
buffer specific lines was mapped to the more aptly mnemonic "go to
(d)efinition".

Don't know yet if I should provide an alternative for `ag` not being
present that is as seamlessly as possible, or that a big fat warning
message should be shown.