Moving to nvim config directory
In moving to Neovim (and its corresponding configuration directory
structure) we get rid of vimrc and make it init.vim.  Therein we change
the directory structure to that of Neovim to make sure the plugins load.

There is more stuff to do, but this change is important enough to
warrant a commit on its own.

Closes #53