Add open line without going to insert mode
When running CLI (which I'm doing since moving to Neovim), mapping a
modifier key with `<CR>` doesn't work, it just moves the cursor down.
This is pissing me off for a while now, so I can't be bothered; at least
with this change it's easy to add a blank line down.

Note that this makes <CR> not work as selection in, for example, the
Quick Fix window!  Use `o` (mnemonic: Open) for that.