Remove UltiSnips
Snippets have been ok, but the time gained is not really relevant IMO.
It's good for boiler plating, but this can be solved in other ways
(scaffolding scripts, for instance).  For the tiny stuff, I'm looking
more to AST parsers and LSP clients.
As a side effect, I'm hoping it might help me type more and thus improve
my typing accuracy and speed... this might be naive ;-)

I'm suspecting that it'll be frustrating sometimes, but given that I
removed the engrained `<TAB>` trigger in [e8060ca] and didn't really
miss it in the last days, I'm thinking it'll not be that hard.