Rhubarb is removed
It doesn't happen a lot that I remove a Tim Pope plugin, but Rhubarb was
mostly used as gimmick, as autocomplete for issue numbers.  Since my
browser has the GitHub issue that I'm currently working open in a tab, I
can easily take that info from there.  It does not take any longer, but
yes, I might make a typo... But this does not warrant a plugin ATM.