Disable nvim-treesitter's highlighting
With the highlighting enabled, indentation is screwed in different ways.
The most apparent is with PHP files when all of a sudden _no_
indentation is done at all.  At first I was thinking this might have to
do with updating to the latest nightly, but after some trial and error,
it became that _something_ in nvim-treesitter highlight module is
causing it.  When disabled, indentation works just fine.
Using the indentation module of treesitter (which I don't by default)
_somewhat_ seemed to resolve this, but when reformatting  etc... this
screwed the indents also in weird ways (some indents were gone, but
others not, was losing much time to go down a rabbit hole and try to
pinpoint it).

I found https://github.com/nvim-treesitter/nvim-treesitter/issues/1136
which seems like a related issue, so I'll keep an eye on that.

It is stated in nvin-treesitter README that it is still very
experimental, so not really bothered by it at all, but I need because of
Playground, which has been a great help in bootstrapping kokwame.