Remove coc#status from status line; screws up screenredraw
When this status is added to light line, it happens that with showing
warnings, the screen redraw is like off by one or something, causing
lines to shift up or down while moving the cursor.
Also, the command line's `:` is rendered in the same color as the mode
in the status line, and in some cases, the status line is rendered
twice.
All in all, this is just too screwed up to be working comfortably.
Since the gutter shows the warnings already, this is enough for me and I
can get by with `:CocStatus` if need be.