Kokwame plugin is managed locally
Adding Kokwame to the runtime path from the locally checked out
repository makes a lot more sense since it is still WIP and changing too
frequently to have Plug manage it.
And, of course, it makes it a lot easier to user test things out before
committing (let alone push it) to the repository.